requirements.txt
python-telegram-bot==20.7
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
psycopg[binary,pool]==3.1.18
pydantic==2.6.4
python-dotenv==1.0.1
httpx==0.25.2
redis==5.0.3
pydantic-settings==2.2.1
python-dateutil==2.8.2
python-decouple==3.8
PyPDF2==3.0.1
langchain==0.1.16
orjson==3.9.15
structlog==24.1.0
loguru==0.7.2
beautifulsoup4==4.12.3
lxml==5.1.0
pytz==2024.1
numpy==1.26.4
scipy==1.12.0
asyncpg==0.29.0
ruff==0.3.5
pytest==8.1.1
aiosqlite==0.19.0
